Removes builder traits
